flatpak: use prebuilt TDLib library
TDLib is not really easy to compile on a normal system: my laptop with 8GB of ram + 8GB of swap (zram) runs out of memory when trying to build it. I decided that bundling tdlib as prebuilt is the best decision to also make it easier to contribute to Telegrand.