Add vscode setting to handle meson project
This commit adds the minimum required settings to work with telegrand in
vscode and derivates in a more convenient way (as an alternative to
Gnome Builder).
First, the commit defines tasks for setting up and installing meson
development builds. It then adds a launch configuration that is based on
these tasks. This can be used together with the extension
https://github.com/vadimcn/vscode-lldb to debug the application.