login: Implement functionality to resend auth code
With this commit, the user can have the authorization code sent to him
in another form (SMS, Call, Flash Call). The user is asked to wait for
the time period suggested by tdlib, but can also resend the code be-
fore it expires.

There is a small workaround:

Sometimes the user may get a FLOOD_WAIT when he/she wants to resend the
authorization code. But then tdlib blocks the resend function for the
user, but does not inform us about it by sending an
'AuthorizationState::WaitCode'. Consequently, the user interface would
still indicate that we are allowed to resend the code. However, we al-
ways get code 8 when we try, indicating that resending does not work.
In this case, we automatically disable the resend feature.