Set the correct system language code for tdlib
While working on #89, I noticed that the ToS text is shown in English
instead of my system language, German.
The reason is that the user's current locale for tdlib is always set
to 'en-US'.
With this commit, the user's locale is set for tdlib. The database
directory needs to be deleted in order to apply the new settings (at
least when the user is still in the login flow).