Implement registration page
This commit just adds a simple page for user registration.

It also brings a Terms of Service (ToS) dialog, that the user can open
via a clickable label. If Telegram enforces the user to accept the ToS,
the login flow will not proceed until the user has done this. Then,
the dalog is forced upon the user when she/he wants to proceed.

What is still missing is a photo upload function, as found in the
official client.