Move transition to phone number page into match arm
We will be called with 'AuthorizationState::WaitPhoneNumber' anyway,
so moving it into the match arm makes it more consistent with the
other pages.
This is a preparation for a commit to implement focus grabbing for
relevant entries after the transition from a previous page. Having
everything squashed into the match arms (transition and grabbing)
will make the code clearer.