WIP: Polish logout behaviour
Provide the user with more feedback by showing an info bar at the
bottom of the window. This infobar contains a spinner and a label
telling the user that she/he is currently being logged out.

Also, to prevent the user to enable the logout action again, set
it disabled after it has been activated for the first time.