Reset error labels and entries in login flow
Currently, when the user makes an invalid input, an error label is
displayed. Now the user corrects the input, presses "Next" and is
redirected to the next page. If the user now presses the "Previous"
button, he is redirected back to the first page and all error labels
are still there, although the input that is there is correct.

The reverse case is just as true: for example, a user enters an
incorrect code, clicks on "Previous", enters the phone number
correctly again and is redirected to the code page where the user
can still see the old error message.

Furthermore, previously made entries are not deleted if the user
clicks on "Previous", so that she/he sees them again in the process of
logging in and they may no longer be related to the inputs that the
user now has to make.

With this commit, it's ensured that the error labels and entries will
always be reset.