login: Revise function 'set_visible_page_name'
First, the function is renamed to 'navigate_to_page', which makes it
much more clear that the function does more than just set the visible
page of the leaflet.

Furthermore, both the error label whose text is to be deleted and the
widget that is to be focused become optional.

This is a preparation for the fact that in the future not every page in
the login flow will have an error label or a widget to be focused (e.g.
password-forgot-page).