login: Revise function 'handle_user_result'
In this commit, the internal routine of error handling is outsourced to
a new function, named 'handle_user_error'. This new function can be
used if only certain errors require the routine.

This is a preparation for the implementation of the password recovery
function. The outsourced function is to be used when the recovery code
is to be sent from the user to tdlib. The error procedure is then
needed separately when an error other than "PASSWORD_RECOVERY_EXPIRED"
occurs.