login: Revise action handing
In this commit, the handling of actions as well as the showing and
hiding of the 'Previous' and 'Next' button is fundamentally revised.
The idea behind this is to enable actions only when they make sense.
In addition, the 'Next' button, similar to the 'Previous' button now,
will no longer be visible on every page in the future.

First of all, a function called 'disable_actions' is implemented to
disable all actions of the login template. This is used in the 'freeze'
function, among others.

Next, a function called 'update_actions_for_visible_page' is implement-
ed. In this function, based on the currently visible page of the
leaflet, all actions are enabled or disabled and the visibility of the
'Previous' and 'Next' button is changed. Among other things, this func-
tion is called in the 'connect_visible_child_name_notify' of the
leaflet and in the error routine of the 'handle_user_error' function.

This commit serves as preparation for the implementation of the
password recovery function, which, as mentioned at the beginning, will
introduce actions that only make sense on certain pages and will hide
the 'Next' button on certain pages.