login: Implement password hint
With this commit, the password hint that the user provided when creat-
ing their cloud password will be displayed on the 'password-page'. The
hint is in a label which is the child in an AdwActionRow below the
password input field.

  +----------------------------------------------------+
  | +------------------------------------------------+ |
  | | Password                                       | |
  | +------------------------------------------------+ |
  +----------------------------------------------------+
  |                                                    |
  | Hint                                     Some Hint |
  |                                                    |
  +----------------------------------------------------+

If the user did not specify a hint, the entire AdwActionRow is hidden.