login: Implement wait-other-device-confirmation
When this commit is applied, the user gets the option to log in via a QR
code, which can be scanned using another device connected to Telegram.

Unlike in the official Telegram client, the function to scan a QR code
is not offered to the user by default when first launched. Instead, the
user has the option to go from the 'welcome page' to a new page where
the QR code will be displayed.