login: Fix actions not enabled after logout
The actions on the 'phone-number-page' are not activated as soon as you
start Telegrand when you are already logged in and then log out again.

The reason for this is that in the match arm 'AuthorizationState::Ready'
all login actions are disabled. However, since the leafet has the
'phone-number-page' as its visible child by default,
'connect_visible_child_name_notify' is never fired when you log out.
Consequently, the actions are then never updated.

With this commit, all login actions are disabled by default when the
'Login' object is instantiated. It also checks in the match arm
'AuthorizationState::WaitPhoneNumber' if the actions need to be
manually updated when there is no transition.