chat-history: Introduce MessageIndicators and MessageLabel
This introduces a new MessageIndicators widget that will contain all the icons and labels that are related to the status of the message (sent/unread icons, timestamp label, edited label, etc), currently only the timestamp is included. This also introduces the MessageLabel which is a widget that contains the label of the related message and also the MessageIndicators widget, which is always placed at the bottom right of the label (the label will always leave a space for it to fit).