session: Fetch own user
If we know which user we are, many options are open to us: For example,
we can hide the online status for ourselves in the sidebar. We can also
give our own chat a different name. Last but not least, in the future we
can always mark the message status in our own chat as 'read' for all
messages.