sidebar-chat-row: Show online status for bilateral chats
The online status is represented by a small dot with the color 'green_4'
above the avatar. To distinguish it well from the avatar, a piece of the
avatar is cut out around the indicator using an OpenGL mask shader. If
for some reason shaders aren't supported on the target system,
'@window_bg_color' is used as a fallback.

The online status is not displayed for both the own chat and chats with
bots.