perf: Use binary search to remove message
The elements in the message item list are ordered. While the day di-
viders are ordered only by their date time, the messages are additional-
ly sorted by their id.

We can exploit this by applying a binary search when looking for the in-
dex of the message to be removed. This leaves us with only a logarithmic
complexity instead of a linear one.