misc: Simplify usage of IndexMap in some places
The use of the IndexMap can be simplified in the 'ChatList::get_chat'
and 'UserList::get_or_create_user' functions by not determining the in-
dex of the element to be found, but finding the element directly.

Additionally, in the 'UserList::get_or_create_user' function the number
of lookups can be reduced from 3 (user is not initially existing) to 1
by using the Entry API and omitting recursion.