cli: Implement basic cli handling
For this purpose, external libraries such as 'clap' are intentionally
omitted. On the one hand more complex functions like subcommands are
not needed and on the other hand clap interferes with the gtk specific
command line handling. Instead, command line options are added to the
gio::Application.

In this first commit to handle command line parameters, the version
number of Telegrand can already be printed.