cli: Set log level using parameter
At the moment the log level can only be influenced with the environment
variable 'RUST_LOG'. This is very specific for the log implementation
'pretty_env_logger' and would not work anymore if another log implemen-
tation would be used in the future. In addition, this is not very ex-
ploratory for the user.

With this commit, the log level can be set as a command line option. On
the one hand this makes the options more explorative for the user. On
the other hand, the option serves as a kind of interface that can be
used to integrate other log implementations in the future.

If the user does not specify a log level, 'debug' is used by default
for debug builds and 'info' for release builds.
On top of that, the log level is also transmitted to tdlib.