cli: Set data dir by parameter for debug builds
Especially for the development of Telegrand it makes sense to specify
another data directory via command line arguments. Because often you
work at different places in the application at the same time (login vs.
inapp) and so you would have the possibility to switch back and forth
quickly.

This option for alternative data directories is only enabled for debug
builds, because it would make little sense to provide them in releases.