sidebar: Enable searching for own chat
This commit allows to search for the own chat in the sidebar. To find
the own chat, the query string must be contained in the string "Saved
Messages" (translated into the respective local language).

This commit is the starting point for treating the own chat differently.
Further commits will follow.