session: Allow more ways to navigate back to the chat list
The user does not necessarily have to use the back button in the header
bar to get back to the chat list in compact mode. Alternatively, the us-
er could for example use the mouse back button or a swipe gesture.

In addition, this commit also changes the behavior of the back button in
the header bar: The chat is deselected only when the leaflet transition
is over.

fixes #173