session: Create own user directly from td user object
This commit serves as preparation for the soon coming multi-account
feature:

It turned out that the update for the own user does not arrive after the
user has logged into an account for the first time. My guess is that
tdlib doesn't send the (redundant) update for the own user anymore if it
has registered that the own user was queried with `GetMe` before.

However, when the application is started again, the update for the own
user arrives. My guess is that the update now arrives before the `GetMe`
function is called.

This commit should ensure that all necessary information about the own
user is available at any time. Therefore the own user is now created di-
rectly from the tdlib user object.