session: Implement multi account feature
The Matrix client Fractal served as a template for this feature.
It is worth mentioning that there is now a new widget, the `Session-
Manager`, which is responsible for managing the sessions. This widget
can be considered as the main view. It is directly subordinate to the
application window. The following sections describe what this commit
does.

Adding New Sessions
------------------------------------------------------------------------
The SessionManager directs it to the Login widget as soon as the user
wants to add a new session. Login returns control to the SessionManager
if either the new session is ready, the login process has been aborted
by the user, or it is noticed that the phone number is already logged
in. Session duplicates caused by QR code logins are handled by tdlib di-
rectly. For that, we pass a list of already logged in user ids to tdlib.

  The names of new database directories follow the scheme `db<t>`, where
`t` is the time in milliseconds that has passed since the Unix epoch.
Thus, sessions are sorted by time according to their creation. The sort-
ing is also used in the application when the sessions are displayed to
the user. This scheme ensures that `db0` will always remain the first
session.

Adding Existing Sessions
------------------------------------------------------------------------
The SessionManager analyzes the individual database directories in the
Telegrand data directory at startup to see which sessions can be logged
in directly. To do this, it checks the presence of a `td.binlog` or a
`td_test.binlog` file.

Destroying sessions
------------------------------------------------------------------------
This is realized by first logging out the client and then deleting the
database directory once the `AuthorizationState::Closed` event has been
received for that session.

  Destroying sessions happens in different places: When the login is
canceled, When the QR code is canceled, when a logged in session is log-
ged out, and when the session is removed from another device.

Remembering Recently Used Sessions
------------------------------------------------------------------------
In order to remember the order in which the user selected the sessions,
the SessionManager uses a gsettings key value pair.

Test Data Center Handling
------------------------------------------------------------------------
The option on the login page to use the test DC has been removed com-
pletely, because this option was global, but the flag to use the Test DC
applies to the account. It therefore no longer made sense to keep it.

  By default, the production server is always used at the first login.
Further accounts can then be connected on the test DC (by long pressing
on "Add Account"). However, to use the test DC at the first login, Tele-
grand can be started with the command line flag `--test-dc` that has
been added in this commit.

Changing Data Directory per Command Line Parameter
------------------------------------------------------------------------
Commit e07a488ec53cd357e2adb402c892e3784fb48c7e introduced a cli parame-
ter to quickly test things with other accounts without having to manual-
ly move the data directory. This parameter has been removed as its pur-
pose becomes obsolete in a multi-account capable application.

Positive Side Effects
------------------------------------------------------------------------
Due to the newly implemented session handling, the logout of a session
works without a visual delay (#63) and information about the currently
logged in user is shown in a menu (#82). Additionally, exiting the QR
code page now works immediately, which could take some time before.

Should also fix #63, #82