sidebar-avatar: Fix incorrect online status
Just unwatching bindings is not always sufficient. If the previous bind-
ing was conditional and this condition no longer applies to the new mod-
el, the property that was last set by this binding is retained.

This is exactly what happened with the online status in the sidebar and
should be fixed with this commit by resetting the property if the origi-
nal condition no longer applies.