session: Fix SyncSender panic
It is indeed possible that the object that created the sender and re-
ceiver and attached it to the default main context has been disposed in
the meantime.

This is problematic if it is now tried to upgrade a weak reference of
this object in the receiver closure. It will either panic directly if
`@default-panic` is used or the sender will return an error in the
`SyncSender::send()` function if `default-return glib::Continue(false)`
is used. In the latter case, the Receiver will be detached from the main
context, which will cause the sending to fail.

To fix this, the result of each send is checked instead of unwrapping.
Only those senders are then kept that were successful.