(WIP) session: Await functions::GetMe before proceeding
It's better to wait for `functions::GetMe` to return our own user be-
fore creating a new session. In this way we don't need expressions for
watching the `me` property to become `Some`. We can just use
`Session::me()`, which at this point guarantees to return our own user.

Since updates are already coming in the time that the own user is being
waited for, `WaitingForOwnUser` is introduced as the fourth variant of
the `ClientInfo` enum. Here all updates for the sessions are cached and
then applied as soon as the own user is available with which the session
can be created.