session: Set own user before showing the session
This commit ensures that `Session::me()` always returns the own user
after the session has been added to the widget tree. This is the second
attempt of a pragmatic approach that prevents that a lot of things are
going to become too complicated. Especially, handling the chat history
for the own user would be a real nightmare.