login: Improve phone number input
This commit improves the phone number input during the login process.
The improvements include the following:

- Load the list of country infos from TDLib
- Add a specialized country info to that list in case the test data
  center is used
- Show a country selection combobox row above the phone number entry
- Update the calling code in the phone number entry when a country is
  selected from the combobox
- Update the selected country info in the combobox when the user changes
  the phone number within the entry
- Highlight the country calling code in the phone number entry
- Keep the highligted country code in the entry intact even when white-
  spaces are inserted
- Select the number except calling code in the phone number entry in
  case of an error
- Determine the system country code and initially set the corresponding
  calling code

The phone number is not formatted during input. The reason for this is
that this de facto only works with whitespaces. To realize this is very
complex and works in certain cases only insufficiently.

Even after the phone number has been confirmed with `Next`, the phone
number will not be formatted. This would make little sense: In some case
the phone number is correct is and the user doesn`t see it after confir-
mation anyway. In the other case, if the phone number is incorrect, it
cannot be formatted in most cases.

fixes #117