misc: Restrict access to inner types of boxed monads
This commit will prevent direct access from outside to the inner type of
a boxed monad. Instead, the `Deref` and `From` trait is implemented.

The `Deref` trait makes it a bit easier to use, as the inner type can
now be accessed directly with `x.y()` instead of `x.0.y()` via the tuple
operator.

The `From` trait is implemented as `impl From<$Boxed> for $Type`. This
makes it possiple to move out the inner type.