ci-cd: Use nightly toolchain for rustfmt
Our style for imports has become very inconsistent. Rustfmt has a set-
ting `imports_granularity` which can unify the style. Unfortunately,
this function is unstable and therefore only available in the nightly
toolchain. Since I fear that this setting will not become stable in the
near future, I suggest we use the nightly toolchain for rustfmt.

In addition, I suggest that we use the value "Module" for this setting.
On the one hand, too deep nesting would, from experience, lead to
greater difficulties in merging. On the other hand, an un-nested import
list unnecessarily bloats the code. I think "Module" is a good trade-
off.

This commit introduces all necessary changes: It creates a
`.rustfmt.toml` file that contains the settings, adapts the CI-CD pipe-
line and unifies the imports.