avatar: Improve separation between widget and data
This commit massively refactors the avatars by improving the
separation between the avatar component (widget) and item (gobject).
The avatar item is now just a simple boxed type around
tdlib's ProfilePhoto or ChatPhotoInfo, so it will no longer store
the paintable and other widget-related properties. The widget,
instead, has become the one that holds all the properties and
expressions to correctly display the avatar of either an User
or a Chat.