message-row: Turn it to an actual widget
The MessageRow is no longer a subclassable type. Now it's just a
"container" widget that can still show the avatar on a side. Now every
message widget is just a GtkWidget subclass, which it means that they
will have custom properties instead of sharing the generic message
GObject, which is not always useful for certain message types.