chat-history: Show a "scroll to bottom" button
In addition, it also scrolls to new messages when one is at the end of
the history.

With autoscroll there is a short flicker when you send a message. For
incoming messages, The flicker is due to the fact that we remove the
pending message once it is send. We should fix this in the future.

I think that the autoscrolling is still an improvement, because at least
now the user doesn't have to scroll all the time.

The commit is strongly inspired by Fractal.

Fixes #29.

Co-authored-by: Marco Melorio <marco.melorio@protonmail.com>