Chat-history: Mark visible messages as viewed
This includes both the bottommost regular message and sponsored
messages.

A regular message is considered as being viewed if parts of it are
either visible inside the specified list view or if the user needs to
scroll up in order to see it.

Inside a list view `L` a sponsored message `S` is considered as being
viewed if `L ∩ S = S` applies.