Message-row: Reduce width of outgoing messages
Currently, two messages with the same content can have different
dimensions depending on whether the message is outgoing or incoming. An
outgoing message has a potentially larger width because it has no avatar
and the space of the missing avatar is directly taken by the message.
This is especially noticeable when the application window has a small
width.

This behavior not only looks very inharmonious but also interferes with
the quick differentiation between incoming and outgoing messages.
Especially pictures that have only a very narrow frame with the accent
color can be problematic.

This is easily solved with this commit, in which the left margin of the
outgoing message is extended by the right margin of an incoming message
(doubling).