chat: Stop adding last messages to ChatHistory
This is wrong for two reasons:

1. The ChatHistory is not a generic list of messages, it's just a list
that shows messages to the user and loads them lazily while needed, so
it's not really useful filling it with all the messages we get.

2. Using the chat last message as the oldest message in ChatHistory is
not really how chat history API should be used in TDLib, so it could
happen that we load the wrong messages in the ChatHistory because
meanwhile the chat last message is updating.