sidebar-search: Reimplement search in this new widget
Compared to the previous implementation, this new widget is completely
separated from the chat list in the sidebar, meaning that we don't use
the overly complicated merged ChatList + UserList model for the ListView
and the same SidebarRow that we already use for the regular (and already
complex) chat list in the Sidebar.

Also, the new widget is more easily expandable to support more object
types, like messages and custom objects (e.g. section headers).

This new widget also gained the ability to show recent searches when the
query is empty.

Part of the new search logic is inspired by Telegram X. A mention about
this will be added in the next commit.