chat-history: Add MessageBubble widget
This widget shares a lot of logic and styles that were scattared ascross
the other message widgets. It will take care of setting and updating the
sender label, it can have a prefix widget and it will take care of the
indicators position based on the presence of text. The widget also sizes
itself with a max width, so that the bubbles doesn't take too much space
when the view is wide.

This also contains some style changes, so that it's more on par with
other clients.