tdlib: Add ChatListItem as ChatList items instead of the Chat
This is a set of fairly big changes that will allow for future support
of chat folders and archived chats.

There's a new `ChatListItem` that contains the `order` and `is-pinned`
state for a `Chat` in a ChatList. This means that `Chat` doesn't contain
these properties anymore. Also, the `ChatList` will not cache all the
chats returned from TDLib anymore. The `Session` will take care of that
instead.

The other big change is that the `ChatList` will take care of both
removing no-more-needed chats from the `Sidebar`'s `GtkListView` and of
keeping the correct chat order. This means that we don't need to use
both a `GtkFilterListModel` and a `GtkSortListModel` anymore.

Co-authored-by: Marcus Behrendt <marcus.behrendt.86@gmail.com>