session-manager: Remove workaround for waiting the own user
This was previously added because for some reason TDLib wasn't sending
the User update in time before returning the response of `get_me`. Since
this is something that is guaranteed to happen by TDLib (confirmed by a
TDLib dev), I'm removing the workaround. If this happens again, it needs
to be investigated and fixed in TDLib itself.