misc: Use GtkInscription where it makes sense
This widget is a lighter alternative to GtkLabel, but with the
limitation that it will not size itself automatically. This is
especially useful in lists, where you often already have a pre-defined
area where you want to place the widgets.

Unfortunately this widget cannot be used in the chat history at all,
because the widgets inside it are actually sized based on the text
inside them.