sidebar-row: Start porting to glib's SignalGroup
Just replace the plain signals for now.

`SidebarRow` code was probably one of the most complicated widgets in
Telegrand, both because there are a lot of bindings needed to ensure we
update everything inside it correctly and also because we used GTK's
expressions to make this bindings, which was ultimately not a good
choice. Even though they are a fairly good addition to GTK for more
declarative UIs, they are not powerful enough for more complex
scenarios, and `SidebarRow` is full of these, and they only made the
code much more complex than it should be.

Glib's new `SignalGroup` is an object that can hold signals of an object
and it automatically disconnects them on dispose. This is a different
approach for binding stuff compared to GTK's expressions (they are not
really declarative), but at least they are more powerful so in the end
they are much better for complex scenarios like `SidebarRow`.