misc: Update gtk-rs and libadwaita-rs
This is a bit more paintful because to do this we are also forced to use
libadwaita 1.3 beta right away, because some 1.2 APIs were removed in
libadwaita-rs and then readded in the 1.3 feature in a different way.
Because of this, this commit also bumps glib, gio, and gtk4 to the
required versions to build libadwaita 1.3.beta.