WIP: chat-history: Move the message list in a separate widget
This a new major change to the chat history. We now have a new widget
component, called MessageListView, which can display messages... in a
list view. For now it's basically the same as the internal list view in
ChatHistory, but in the future it will support showing messages from
other contexts, like pinned messages history, reply threads, post
comments, etc.