feat(ui): Add component `MapMarker`
This component can be used as a marker on a `shumate::Map` and display
any widget. The component has a special layouting which causes the
component to be twice as high as its content. The content is always in
the upper half. As a result, the bottom edge of the set widget of the
component ends exactly at the geo position of the map.