fix(ui/chat-history): Load sufficient number of messages
Up to now, we have only ever loaded a fixed number of messages
initially. This number may be too few, so that the scrollbar of the
ChatHistory does not appear. Thus, the user must first resize the
application to load older messages.

This commit ensures that initially just enough messages are loaded so
that the scrollbar appears.